requirements.txt
# Core
livekit-agents[openai,deepgram,cartesia,silero]==1.2.14
livekit-plugins-anthropic
livekit-plugins-assemblyai
livekit-plugins-elevenlabs
livekit-plugins-groq
python-dotenv

# Type safety and templating
pydantic>=2.0.0
jinja2>=3.1.0

# Memory and storage
aiosqlite>=0.19.0

# Code quality
ruff>=0.1.0
black>=23.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pre-commit>=3.0.0   